requirements.txt
arrow==1.3.0
attrs==23.2.0
bravado==11.0.3
bravado-core==6.1.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
ciso8601==2.3.1
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
discord-webhook==1.1.0
fonttools==4.52.1
fqdn==1.5.1
future==1.0.0
git-filter-repo==2.38.0
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
hyperopt==0.2.7
idna==3.7
influxdb-client==1.43.0
isoduration==20.11.0
joblib==1.4.2
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
matplotlib==3.7.1
monotonic==1.6
msgpack==1.0.8
networkx==3.3
numpy==1.25.0
oauth2client==4.1.3
packaging==24.0
pandas==2.0.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
py4j==0.10.9.7
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybit==5.7.0
pycryptodome==3.20.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyti==0.2.2
pytz==2024.1
PyYAML==6.0.1
reactivex==4.0.4
referencing==0.35.1
requests==2.32.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.0
simplejson==3.19.2
six==1.16.0
swagger-spec-validator==3.0.3
TA-Lib==0.4.26
threadpoolctl==3.5.0
tqdm==4.66.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.0
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
webcolors==1.13
websocket-client==1.8.0
websockets==12.0
